Fix: 게시글 권한 설정 및 유저와 댓글/게시글 연관관계 반영

+++ board/src/main/java/com/example/board/article/service/ArticleService.java
@@ -6,22 +6,28 @@ import com.example.board.article.repository.ArticleRepository;
 import com.example.board.comment.model.dto.CommentDetailDto;
 import com.example.board.comment.model.entity.Comment;
 import com.example.board.comment.repository.CommentRepository;
+import com.example.board.user.model.entity.User;
+import com.example.board.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
 public class ArticleService {
     private final ArticleRepository articleRepository;
     private final CommentRepository commentRepository;
+    private final UserService userService;
 
     @Transactional
     public Page<ArticleSummaryDto> getArticleList(int page) {
@@ -36,7 +42,7 @@ public class ArticleService {
                             article.getCommentCount(),
                             article.getLikeCount(),
                             article.getCreatedAt(),
-                            null
+                            article.getUser().getNickName()
                     )
             );
         }
@@ -45,16 +51,19 @@ public class ArticleService {
 
     @Transactional
     public CreateArticleResDto createArticle(CreateArticleReqDto reqDto) {
+        User loginUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         Article article = new Article(
                 reqDto.getTitle(),
-                reqDto.getContent()
+                reqDto.getContent(),
+                loginUser
         );
         articleRepository.save(article);
         return new CreateArticleResDto(
                 article.getId(),
                 article.getTitle(),
                 article.getContent(),
-                article.getCreatedAt()
+                article.getCreatedAt(),
+                article.getUser().getNickName()
         );
     }
 
@@ -87,14 +96,24 @@ public class ArticleService {
     @Transactional
     public UpdateArticleResDto updateArticle(Long id, UpdateArticleReqDto reqDto) {
         Article article = articleRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Article not found"));
+        String loginUserId = userService.getLoginUserInfo().getUserId();
+        String authorUserId = article.getUser().getUserId();
+        if (!Objects.equals(loginUserId, authorUserId)) {
+            throw new AccessDeniedException("Login user does not match author");
+        }
         article.update(reqDto.getTitle(), reqDto.getContent());
         articleRepository.save(article);
         return new UpdateArticleResDto(article);
     }
 
-    // TODO: 게시글 삭제 api 완성할 것
     @Transactional
     public void deleteArticle(Long id) {
+        Article article = articleRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Article Not Found"));
+        String loginUserId = userService.getLoginUserInfo().getUserId();
+        String authorUserId = article.getUser().getUserId();
+        if (!Objects.equals(loginUserId, authorUserId)) {
+            throw new AccessDeniedException("Login user does not match author");
+        }
         articleRepository.deleteById(id);
     }
 }
